tests: move logger tests to different module
